moved color channel value check to check_functions.py / added make_error_msg to __init__.py
